The following message related to the next commit.
This commit is a snapshot of the working environment before I
start changing the Lexer.

--- Origianal commit message

Prints tokens from a file.

Removed the previous attempts to do what I wanted.
They will remain in the master branch
This is a more or less release official branch.
At lest this branch will be exported to the formatter.
Huge refactoring.