This commit removes the usages of diag:: namespace, which is in general conflicting with llvm itself when I try to use this project as a dependency in clang.
libsourcekit
  The problem is that this library relies on LLVMSupport.a
library that has global definitions of command line
options in it, and an assertion that checks that no
duplicate options have been passed. But of course they
are passed as the global scope (with all the variable
declarations) is processed twice.

UPD.: This project cannot be standalone in principle.
It is too difficult to make it standalone and use it
in Clang, as was originally intended. This project
relies on a modified version of llvm source tree, and
it cannot work in principle because of conflicts.

Instead, one should work insdie the Swift source tree and
modify its own llvm tree if needed.